Update network.md
I have reworked the wording of the questions and answers for clarity, grammar, spelling, and order. I recommend we remove the following questions (have not reworded these):

--What ports must be open on a node?
--How can I open my ports on my node machine? Can you describe it for Mac and Linux?
--Is it possible to check if your validator is running correctly and what the rewards are without using Grafana/Prometheus?

These three are per Fabian not wanting commands in the FAQ. The step-by-step node setup guide should cover them.

--Is there an advantage to splitting validators across multiple nodes instead of running all validators on one node?

This question was asked when the community was under the assumption of a 32 LYX min balance, when (in their minds) running hundreds of validators was a realistic scenario.

--How should the logs look in each case to have certainty everything runs smoothly?
I think it would be better to have a section in the l16 guide for logs, where common errors (and solutions) are listed.